fix: show TLS certificates for bare hostnames
