+++ src/shared/index.ts
@@ -2,4 +2,5 @@ export * from "./responseSchemas.js";
 export * from "./Service.js";
 export * from "./serviceCapabilities.js";
 export * from "./ServiceWithPosition.js";
+export * from "./tlsEndpoint.js";
 export * from "./types.js";
